basic content display in srs + more flashes in srs

+++ src/routes/(protected)/study/[id]/+page.server.ts
@@ -17,6 +17,7 @@ import {
 } from '$lib/config/zod-schemas';
 import { getTodayCount } from '$lib/server/database/models/study';
 import { isUUID } from '$lib/_helpers/isUIID';
+import { setFlash } from 'sveltekit-flash-message/server';
 
 //ограничить для очереди дублируем опять
 const countCardsByState = (cards: StudyCard[]): Count => {
@@ -63,34 +64,59 @@ export const load = (async (event) => {
 export const actions = {
 	good: async (event) => {
 		const goodForm = await superValidate(event, zod(gradeCardSchema));
-		const res = await gradeStudyCard(goodForm.data.studyCardId, 'Good');
-		console.log(res);
-		console.log('remembe)');
-		return { goodForm };
+		if (!goodForm.valid) {
+            setFlash({type: 'error', message: 'Ошибка валидации формы'}, event);
+            return { goodForm };
+        }
+        const res = await gradeStudyCard(goodForm.data.studyCardId, 'Good');
+        if (res) {
+            setFlash({type: 'success', message: 'Карточка успешно оценена как хорошая'}, event); //TODO: время следующего повторения
+        } else {
+            setFlash({type: 'error', message: 'Не удалось оценить карточку как хорошую'}, event);
+        }
+        return { goodForm };
 	},
 	again: async (event) => {
 		const againForm = await superValidate(event, zod(gradeCardSchema));
-		const res = await gradeStudyCard(againForm.data.studyCardId, 'Again');
-		console.log(res);
-		console.log('forget(');
-		return { againForm };
+		if (!againForm.valid) {
+            setFlash({type: 'error', message: 'Ошибка валидации формы'}, event);
+            return { againForm };
+        }
+        const res = await gradeStudyCard(againForm.data.studyCardId, 'Again');
+        if (res) {
+            setFlash({type: 'success', message: 'Карточка успешно оценена как "снова"'}, event);
+        } else {
+            setFlash({type: 'error', message: 'Не удалось оценить карточку как "снова"'}, event);
+        }
+        return { againForm };
 	},
 	settings: async (event) => {
 		const studyDeckId = event.params.id;
 		const form = await superValidate(event, zod(studyDeckSettingsSchema));
-		await setNewLimit(studyDeckId, form.data.limit);
-		return { form };
+		if (!form.valid) {
+            setFlash({type: 'error', message: 'Ошибка валидации формы'}, event);
+            return { form };
+        }
+        await setNewLimit(studyDeckId, form.data.limit);
+        setFlash({type: 'success', message: 'Новый лимит успешно установлен'}, event);
+        return { form };
 	},
 	delete: async (event) => {
 		const studyDeckId = event.params.id;
 		const deletForm = await superValidate(event, zod(deleteDeckSchema));
 		const userId = event.locals.user?.id;
-		if (!userId || !isUUID(studyDeckId)) {
-			return fail(400, {});
-		}
+		if (!userId || !isUUID(studyDeckId) || !deletForm.valid) {
+            setFlash({type: 'error', message: 'Ошибка валидации формы или неверный ID колоды'}, event);
+            return fail(400, {});
+        }
 
-		await deleteStudyDeck(studyDeckId);
+        const result = await deleteStudyDeck(studyDeckId);
+        if (result){
+            setFlash({type: 'success', message: 'Колода успешно удалена'}, event);
+        } else {
+            setFlash({type: 'error', message: 'Не удалось удалить колоду'}, event);
+        }
 
-		redirect(302, '/dashboard');
+        redirect(302, '/dashboard');
 	}
 };
